adding support for rendering video classifications via draw_labels()
