+++ fiftyone/core/collections.py
@@ -5319,9 +5319,8 @@ class SampleCollection(object):
 
         Args:
             output_dir: the directory to write the annotated media
-            label_fields (None): a list of :class:`fiftyone.core.labels.Label`
-                fields to render. By default, all
-                :class:`fiftyone.core.labels.Label` fields are drawn
+            label_fields (None): a list of label fields to render. By default,
+                all :class:`fiftyone.core.labels.Label` fields are drawn
             overwrite (False): whether to delete ``output_dir`` if it exists
                 before rendering
             config (None): an optional
@@ -5341,16 +5340,13 @@ class SampleCollection(object):
                     output_dir,
                 )
 
-        if self.media_type == fom.VIDEO:
-            if label_fields is None:
-                label_fields = _get_frame_label_fields(self)
-
-                return foua.draw_labeled_videos(
-                    self, output_dir, label_fields=label_fields, config=config
-                )
-
         if label_fields is None:
-            label_fields = _get_image_label_fields(self)
+            label_fields = self._get_label_fields()
+
+        if self.media_type == fom.VIDEO:
+            return foua.draw_labeled_videos(
+                self, output_dir, label_fields=label_fields, config=config
+            )
 
         return foua.draw_labeled_images(
             self, output_dir, label_fields=label_fields, config=config
@@ -6866,20 +6862,6 @@ def _get_matching_fields(sample_collection, patt, frames=False):
         schema = sample_collection.get_field_schema()
 
     return fnmatch.filter(list(schema.keys()), patt)
-
-
-def _get_image_label_fields(sample_collection):
-    label_fields = sample_collection.get_field_schema(
-        ftype=fof.EmbeddedDocumentField, embedded_doc_type=fol.ImageLabel
-    )
-    return list(label_fields.keys())
-
-
-def _get_frame_label_fields(sample_collection):
-    label_fields = sample_collection.get_frame_field_schema(
-        ftype=fof.EmbeddedDocumentField, embedded_doc_type=fol.ImageLabel
-    )
-    return list(label_fields.keys())
 
 
 def _get_default_label_fields_for_exporter(
